requirements.txt
annotated-types==0.6.0
anyio==4.2.0
assemblyai==0.20.2
certifi==2023.11.17
charset-normalizer==3.3.2
deepl==1.16.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.7.2
PyAudio==0.2.14
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
websockets==12.0
